feat: add the "endo" macro branch
