+++ src/tests/curve.rs
@@ -438,4 +438,31 @@ macro_rules! curve_testing_suite {
             )*
         }
     };
+
+    ($curve: ident, "endo", $z_other_raw: expr) => {
+        #[test]
+        fn test_endo() {
+            use rand_core::OsRng;
+
+            let z_impl = <$curve as CurveExt>::ScalarExt::ZETA;
+            let z_other = <$curve as CurveExt>::ScalarExt::from_raw($z_other_raw);
+
+            assert_eq!(z_impl * z_impl + z_impl, -<$curve as CurveExt>::ScalarExt::ONE);
+            assert_eq!(z_other * z_other + z_other, -<$curve as CurveExt>::ScalarExt::ONE);
+
+            for _ in 0..100000 {
+                let k = <$curve as CurveExt>::ScalarExt::random(OsRng);
+                let (k1, k1_neg, k2, k2_neg) = $curve::decompose_scalar(&k);
+                if k1_neg & k2_neg {
+                    assert_eq!(k, -<$curve as CurveExt>::ScalarExt::from_u128(k1) + <$curve as CurveExt>::ScalarExt::ZETA * <$curve as CurveExt>::ScalarExt::from_u128(k2))
+                } else if k1_neg {
+                    assert_eq!(k, -<$curve as CurveExt>::ScalarExt::from_u128(k1) - <$curve as CurveExt>::ScalarExt::ZETA * <$curve as CurveExt>::ScalarExt::from_u128(k2))
+                } else if k2_neg {
+                    assert_eq!(k, <$curve as CurveExt>::ScalarExt::from_u128(k1) + <$curve as CurveExt>::ScalarExt::ZETA * <$curve as CurveExt>::ScalarExt::from_u128(k2))
+                } else {
+                    assert_eq!(k, <$curve as CurveExt>::ScalarExt::from_u128(k1) - <$curve as CurveExt>::ScalarExt::ZETA * <$curve as CurveExt>::ScalarExt::from_u128(k2))
+                }
+            }
+        }
+    };
 }
